refactor: :recycle: update factory structure
