+++ src/core/drinks/useCases/categories/delete/deleteCategory.container.ts
@@ -1,4 +1,4 @@
-import { resolveCategoriesRepository } from '@/core/drinks/container';
+import { resolveCategoriesRepository } from '@/core/drinks/infra/mongo/container';
 import { DeleteCategoryService } from './DeleteCategory.service';
 
 const categoriesRepository = resolveCategoriesRepository();
