+++ src/core/drinks/useCases/categories/get/getCategory.container.ts
@@ -1,4 +1,4 @@
-import { resolveCategoriesRepository } from '@/core/drinks/container';
+import { resolveCategoriesRepository } from '@/core/drinks/infra/mongo/container';
 import { GetCategoryService } from './GetCategory.service';
 
 const categoriesRepository = resolveCategoriesRepository();
