+++ src/core/drinks/useCases/drinks/create/createDrink.container.ts
@@ -1,4 +1,4 @@
-import { resolveDrinksRepository, resolveIngredientsRepository } from '@/core/drinks/container';
+import { resolveDrinksRepository, resolveIngredientsRepository } from '@/core/drinks/infra/mongo/container';
 import { CreateDrinkService } from './CreateDrink.service';
 
 const drinksRepository = resolveDrinksRepository();
