+++ src/core/drinks/useCases/drinks/create/createDrink.spec.ts
@@ -6,7 +6,9 @@ import { beforeEach, describe, expect, it } from 'vitest';
 import { CreateDrinkService } from './CreateDrink.service';
 import { Ingredient } from '@/core/drinks/entities/ingredient.entity';
 import { BadRequestError } from '@/shared/error/error.lib';
-import { createCategoryFactory, createDrinkFactory, createIngredientFactory } from '@/core/drinks/container';
+import { createCategoryFactory } from '@/core/drinks/factories/category.factories';
+import { createIngredientFactory } from '@/core/drinks/factories/ingredient.factories';
+import { createDrinkFactory } from '@/core/drinks/factories/drink.factories';
 import { IIngredientsRepository } from '@/core/drinks/repositories/IIngredients.repository';
 import { ICategoriesRepository } from '@/core/drinks/repositories/ICategories.repository';
 import { IDrinksRepository } from '@/core/drinks/repositories/IDrinks.repository';
