+++ src/core/drinks/useCases/drinks/delete/deleteDrink.spec.ts
@@ -5,7 +5,9 @@ import { ObjectId } from 'mongodb';
 import { afterAll, beforeEach, describe, expect, it, vi } from 'vitest';
 import { DeleteDrinkService } from './DeleteDrink.service';
 import { BadRequestError } from '@/shared/error/error.lib';
-import { createCategoryFactory, createDrinkFactory, createIngredientFactory } from '@/core/drinks/container';
+import { createCategoryFactory } from '@/core/drinks/factories/category.factories';
+import { createIngredientFactory } from '@/core/drinks/factories/ingredient.factories';
+import { createDrinkFactory } from '@/core/drinks/factories/drink.factories';
 import { ICategoriesRepository } from '@/core/drinks/repositories/ICategories.repository';
 import { IIngredientsRepository } from '@/core/drinks/repositories/IIngredients.repository';
 import { IDrinksRepository } from '@/core/drinks/repositories/IDrinks.repository';
