+++ src/core/drinks/useCases/drinks/get/getDrink.container.ts
@@ -1,4 +1,4 @@
-import { resolveDrinksRepository } from '@/core/drinks/container';
+import { resolveDrinksRepository } from '@/core/drinks/infra/mongo/container';
 import { GetDrinkService } from './GetDrink.service';
 import { resolveStorageProvider } from '@/shared/providers/storage';
