+++ src/core/drinks/useCases/drinks/updateThumbnail/updateDrinkThumbnail.container.ts
@@ -1,4 +1,4 @@
-import { resolveDrinksRepository } from '@/core/drinks/container';
+import { resolveDrinksRepository } from '@/core/drinks/infra/mongo/container';
 import { resolveStorageProvider } from '@/shared/providers/storage';
 import { UpdateDrinkThumbnailService } from './UpdateDrinkThumbnail.service';
