+++ src/core/drinks/useCases/drinks/updateThumbnail/updateDrinkThumbnail.spec.ts
@@ -5,7 +5,9 @@ import { CategoriesRepositoryInMemory } from '@/core/drinks/repositories/inMemor
 import { BadRequestError } from '@/shared/error/error.lib';
 import { ICategoriesRepository } from '@/core/drinks/repositories/ICategories.repository';
 import { IIngredientsRepository } from '@/core/drinks/repositories/IIngredients.repository';
-import { createCategoryFactory, createDrinkFactory, createIngredientFactory } from '@/core/drinks/container';
+import { createCategoryFactory } from '@/core/drinks/factories/category.factories';
+import { createIngredientFactory } from '@/core/drinks/factories/ingredient.factories';
+import { createDrinkFactory } from '@/core/drinks/factories/drink.factories';
 import { Ingredient } from '@/core/drinks/entities/ingredient.entity';
 import { IDrinksRepository } from '@/core/drinks/repositories/IDrinks.repository';
 import { DrinksRepositoryInMemory } from '@/core/drinks/repositories/inMemory/Drinks.repository';
