+++ src/core/drinks/useCases/ingredients/create/createIngredient.container.ts
@@ -1,4 +1,4 @@
-import { resolveCategoriesRepository, resolveIngredientsRepository } from '@/core/drinks/container';
+import { resolveCategoriesRepository, resolveIngredientsRepository } from '@/core/drinks/infra/mongo/container';
 import { CreateIngredientService } from './CreateIngredient.service';
 
 const categoriesRepository = resolveCategoriesRepository();
