+++ src/core/drinks/useCases/ingredients/list/listIngredients.spec.ts
@@ -2,7 +2,8 @@ import { beforeEach, describe, expect, it } from 'vitest';
 import { IngredientsRepositoryInMemory } from '@/core/drinks/repositories/inMemory/Ingredients.repository';
 import { ListIngredientsService } from './ListIngredients.service';
 import { IIngredientsRepository } from '@/core/drinks/repositories/IIngredients.repository';
-import { createCategoryFactory, createIngredientFactory } from '@/core/drinks/container';
+import { createCategoryFactory } from '@/core/drinks/factories/category.factories';
+import { createIngredientFactory } from '@/core/drinks/factories/ingredient.factories';
 import { DEFAULT_QUERY_PARAMS } from '@/shared/constants/query.constants';
 import { ICategoriesRepository } from '@/core/drinks/repositories/ICategories.repository';
 import { CategoriesRepositoryInMemory } from '@/core/drinks/repositories/inMemory/Categories.repository';
