+++ src/core/drinks/useCases/ingredients/update/updateIngredient.spec.ts
@@ -6,7 +6,8 @@ import { CategoriesRepositoryInMemory } from '@/core/drinks/repositories/inMemor
 import { BadRequestError } from '@/shared/error/error.lib';
 import { ICategoriesRepository } from '@/core/drinks/repositories/ICategories.repository';
 import { IIngredientsRepository } from '@/core/drinks/repositories/IIngredients.repository';
-import { createCategoryFactory, createIngredientFactory } from '@/core/drinks/container';
+import { createCategoryFactory } from '@/core/drinks/factories/category.factories';
+import { createIngredientFactory } from '@/core/drinks/factories/ingredient.factories';
 import { Category } from '@/core/drinks/entities/category.entity';
 import { Ingredient } from '@/core/drinks/entities/ingredient.entity';
